+++ src/Game/Client/ClientGameExports.h
@@ -46,7 +46,7 @@ public:
     /**
     *   @brief  For debugging problems when out-of-date entity origin is referenced.
     **/
-    void CheckEntityPresent(int32_t entityNumber, const std::string &what) final;
+    void CheckEntityPresent(int64_t entityNumber, const std::string &what) final;
 #endif
